mv encoding util to consumer (pvtdatastorage)
Currently, reverse order encoding/decoding is present in the
core/ledger/util. Given that it is used only in
pvtdatastorage package and we need to keep such code closer
to the consumer, we move the endcoding/decoding util to
pvtdatastorage package itself.
