+++ core/ledger/pvtdatastorage/kv_encoding.go
@@ -8,12 +8,12 @@ package pvtdatastorage
 
 import (
 	"bytes"
+	"encoding/binary"
 	"math"
 
 	"github.com/golang/protobuf/proto"
 	"github.com/hyperledger/fabric-protos-go/ledger/rwset"
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/version"
-	"github.com/hyperledger/fabric/core/ledger/util"
 	"github.com/pkg/errors"
 	"github.com/willf/bitset"
 )
@@ -109,7 +109,15 @@ func decodeDataValue(datavalueBytes []byte) (*rwset.CollectionPvtReadWriteSet, e
 
 func encodeMissingDataKey(key *missingDataKey) []byte {
 	if key.isEligible {
-		keyBytes := append(eligibleMissingDataKeyPrefix, util.EncodeReverseOrderVarUint64(key.blkNum)...)
+		// When missing pvtData reconciler asks for missing data info,
+		// it is necessary to pass the missing pvtdata info associated with
+		// the most recent block so that missing pvtdata in the state db can
+		// be fixed sooner to reduce the "private data matching public hash version
+		// is not available" error during endorserments. In order to give priority
+		// to missing pvtData in the most recent block, we use reverse order
+		// preserving encoding for the missing data key. This simplifies the
+		// implementation of GetMissingPvtDataInfoForMostRecentBlocks().
+		keyBytes := append(eligibleMissingDataKeyPrefix, encodeReverseOrderVarUint64(key.blkNum)...)
 		keyBytes = append(keyBytes, []byte(key.ns)...)
 		keyBytes = append(keyBytes, nilByte)
 		return append(keyBytes, []byte(key.coll)...)
@@ -119,13 +127,13 @@ func encodeMissingDataKey(key *missingDataKey) []byte {
 	keyBytes = append(keyBytes, nilByte)
 	keyBytes = append(keyBytes, []byte(key.coll)...)
 	keyBytes = append(keyBytes, nilByte)
-	return append(keyBytes, []byte(util.EncodeReverseOrderVarUint64(key.blkNum))...)
+	return append(keyBytes, []byte(encodeReverseOrderVarUint64(key.blkNum))...)
 }
 
 func decodeMissingDataKey(keyBytes []byte) *missingDataKey {
 	key := &missingDataKey{nsCollBlk: nsCollBlk{}}
 	if keyBytes[0] == eligibleMissingDataKeyPrefix[0] {
-		blkNum, numBytesConsumed := util.DecodeReverseOrderVarUint64(keyBytes[1:])
+		blkNum, numBytesConsumed := decodeReverseOrderVarUint64(keyBytes[1:])
 
 		splittedKey := bytes.Split(keyBytes[numBytesConsumed+1:], []byte{nilByte})
 		key.ns = string(splittedKey[0])
@@ -138,7 +146,7 @@ func decodeMissingDataKey(keyBytes []byte) *missingDataKey {
 	splittedKey := bytes.SplitN(keyBytes[1:], []byte{nilByte}, 3) //encoded bytes for blknum may contain empty bytes
 	key.ns = string(splittedKey[0])
 	key.coll = string(splittedKey[1])
-	key.blkNum, _ = util.DecodeReverseOrderVarUint64(splittedKey[2])
+	key.blkNum, _ = decodeReverseOrderVarUint64(splittedKey[2])
 	key.isEligible = false
 	return key
 }
@@ -156,11 +164,11 @@ func decodeMissingDataValue(bitmapBytes []byte) (*bitset.BitSet, error) {
 }
 
 func encodeCollElgKey(blkNum uint64) []byte {
-	return append(collElgKeyPrefix, util.EncodeReverseOrderVarUint64(blkNum)...)
+	return append(collElgKeyPrefix, encodeReverseOrderVarUint64(blkNum)...)
 }
 
 func decodeCollElgKey(b []byte) uint64 {
-	blkNum, _ := util.DecodeReverseOrderVarUint64(b[1:])
+	blkNum, _ := decodeReverseOrderVarUint64(b[1:])
 	return blkNum
 }
 
@@ -177,8 +185,8 @@ func decodeCollElgVal(b []byte) (*CollElgInfo, error) {
 }
 
 func createRangeScanKeysForEligibleMissingDataEntries(blkNum uint64) (startKey, endKey []byte) {
-	startKey = append(eligibleMissingDataKeyPrefix, util.EncodeReverseOrderVarUint64(blkNum)...)
-	endKey = append(eligibleMissingDataKeyPrefix, util.EncodeReverseOrderVarUint64(0)...)
+	startKey = append(eligibleMissingDataKeyPrefix, encodeReverseOrderVarUint64(blkNum)...)
+	endKey = append(eligibleMissingDataKeyPrefix, encodeReverseOrderVarUint64(0)...)
 
 	return startKey, endKey
 }
@@ -211,7 +219,44 @@ func datakeyRange(blockNum uint64) (startKey, endKey []byte) {
 }
 
 func eligibleMissingdatakeyRange(blkNum uint64) (startKey, endKey []byte) {
-	startKey = append(eligibleMissingDataKeyPrefix, util.EncodeReverseOrderVarUint64(blkNum)...)
-	endKey = append(eligibleMissingDataKeyPrefix, util.EncodeReverseOrderVarUint64(blkNum-1)...)
+	startKey = append(eligibleMissingDataKeyPrefix, encodeReverseOrderVarUint64(blkNum)...)
+	endKey = append(eligibleMissingDataKeyPrefix, encodeReverseOrderVarUint64(blkNum-1)...)
 	return
+}
+
+// encodeReverseOrderVarUint64 returns a byte-representation for a uint64 number such that
+// the number is first subtracted from MaxUint64 and then all the leading 0xff bytes
+// are trimmed and replaced by the number of such trimmed bytes. This helps in reducing the size.
+// In the byte order comparison this encoding ensures that EncodeReverseOrderVarUint64(A) > EncodeReverseOrderVarUint64(B),
+// If B > A
+func encodeReverseOrderVarUint64(number uint64) []byte {
+	bytes := make([]byte, 8)
+	binary.BigEndian.PutUint64(bytes, math.MaxUint64-number)
+	numFFBytes := 0
+	for _, b := range bytes {
+		if b != 0xff {
+			break
+		}
+		numFFBytes++
+	}
+	size := 8 - numFFBytes
+	encodedBytes := make([]byte, size+1)
+	encodedBytes[0] = proto.EncodeVarint(uint64(numFFBytes))[0]
+	copy(encodedBytes[1:], bytes[numFFBytes:])
+	return encodedBytes
+}
+
+// decodeReverseOrderVarUint64 decodes the number from the bytes obtained from function 'EncodeReverseOrderVarUint64'.
+// Also, returns the number of bytes that are consumed in the process
+func decodeReverseOrderVarUint64(bytes []byte) (uint64, int) {
+	s, _ := proto.DecodeVarint(bytes)
+	numFFBytes := int(s)
+	decodedBytes := make([]byte, 8)
+	realBytesNum := 8 - numFFBytes
+	copy(decodedBytes[numFFBytes:], bytes[1:realBytesNum+1])
+	numBytesConsumed := realBytesNum + 1
+	for i := 0; i < numFFBytes; i++ {
+		decodedBytes[i] = 0xff
+	}
+	return (math.MaxUint64 - binary.BigEndian.Uint64(decodedBytes)), numBytesConsumed
 }
