--- core/ledger/util/uint64_encoding.go
@@ -1,51 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package util
-
-import (
-	"encoding/binary"
-	"math"
-
-	"github.com/golang/protobuf/proto"
-)
-
-// EncodeReverseOrderVarUint64 returns a byte-representation for a uint64 number such that
-// the number is first subtracted from MaxUint64 and then all the leading 0xff bytes
-// are trimmed and replaced by the number of such trimmed bytes. This helps in reducing the size.
-// In the byte order comparison this encoding ensures that EncodeReverseOrderVarUint64(A) > EncodeReverseOrderVarUint64(B),
-// If B > A
-func EncodeReverseOrderVarUint64(number uint64) []byte {
-	bytes := make([]byte, 8)
-	binary.BigEndian.PutUint64(bytes, math.MaxUint64-number)
-	numFFBytes := 0
-	for _, b := range bytes {
-		if b != 0xff {
-			break
-		}
-		numFFBytes++
-	}
-	size := 8 - numFFBytes
-	encodedBytes := make([]byte, size+1)
-	encodedBytes[0] = proto.EncodeVarint(uint64(numFFBytes))[0]
-	copy(encodedBytes[1:], bytes[numFFBytes:])
-	return encodedBytes
-}
-
-// DecodeReverseOrderVarUint64 decodes the number from the bytes obtained from function 'EncodeReverseOrderVarUint64'.
-// Also, returns the number of bytes that are consumed in the process
-func DecodeReverseOrderVarUint64(bytes []byte) (uint64, int) {
-	s, _ := proto.DecodeVarint(bytes)
-	numFFBytes := int(s)
-	decodedBytes := make([]byte, 8)
-	realBytesNum := 8 - numFFBytes
-	copy(decodedBytes[numFFBytes:], bytes[1:realBytesNum+1])
-	numBytesConsumed := realBytesNum + 1
-	for i := 0; i < numFFBytes; i++ {
-		decodedBytes[i] = 0xff
-	}
-	return (math.MaxUint64 - binary.BigEndian.Uint64(decodedBytes)), numBytesConsumed
-}
